requirements.txt
streamlit==1.32.2
pandas==2.1.4
numpy==1.26.4
scikit-learn==1.3.2
sentence-transformers==2.6.1
torch==2.2.2
transformers==4.38.2
spacy==3.7.4
PyPDF2==3.0.1
pyresparser==1.0.6
streamlit-tags==1.2.8
pillow==10.2.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
requests==2.31.0
python-dotenv==1.0.1
pgvector==0.2.5
pdfminer.six==20221105